+++ lib/turnable_page.dart
@@ -1,9 +1,11 @@
 // Widget
 export 'src/enums/page_view_mode.dart';
 export 'src/enums/flip_corner.dart';
+export 'src/enums/page_flip_event.dart';
 export 'src/widgets/page_flip_controller.dart';
 export 'src/widgets/turnable_page.dart';
 export 'src/flip/flip_settings.dart' show FlipSettings;
 export 'src/model/paper_boundary_decoration.dart';
+export 'src/event/event_object.dart' show WidgetEvent, EventCallback;
 export 'src/widgets/turnable_page_view.dart';
 export 'src/widgets/turnable_pdf.dart';